pyproject.toml
[tool.poetry]
name = "nautobot-plugin-nornir"
version = "0.9.0"
description = "Nautobot Nornir plugin."
authors = ["Network to Code, LLC <opensource@networktocode.com>"]

readme = "README.md"
homepage = "https://github.com/nautobot/nautobot-plugin-nornir"
repository = "https://github.com/nautobot/nautobot-plugin-nornir"
keywords = ["nautobot", "nautobot-plugin"]
include = [

    "README.md",
]
packages = [
    { include = "nautobot_plugin_nornir" },
]

[tool.poetry.dependencies]
python = "^3.6"
nornir-nautobot = "^1.0.0"
nautobot = "^1.0.0-alpha.1"

[tool.poetry.dev-dependencies]
invoke = "^1.4.1"
black = "^20.8b1"
bandit = "^1.6.2"
pylint = "^2.5.2"
pylint-django = "^2.0.15"
pydocstyle = "^5.0.2"
flake8 = "^3.8.3"

[tool.black]
line-length = 120
target-version = ['py36']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | settings.py     # This is where you define files that should not be stylized by black
                     # the root of the project
)
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.pylint.master]
# Include the pylint_django plugin to avoid spurious warnings about Django patterns
load-plugins="pylint_django"

# Don't cache data for later comparisons
persistent="no"

# Don't raise alarms if args/kwargs has an issue, as may be required, just as a decorator
ignored-argument-names="args|kwargs"

[tool.pylint.basic]
# No docstrings required for private methods (Pylint default), or for test_ functions, or for inner Meta classes.
no-docstring-rgx="^(_|test_|Meta$)"

[tool.pylint.messages_control]
# Line length is enforced by Black, so pylint doesn't need to check it.
# Pylint and Black disagree about how to format multi-line arrays; Black wins.
disable = """,
    line-too-long,
    bad-continuation,
    too-few-public-methods,
    duplicate-code,
    """

[tool.pylint.miscellaneous]
# Don't flag TODO as a failure, let us commit with things that still need to be done in the code
notes = """,
    FIXME,
    XXX,
    """